Review CountTable()
Table formatting currently makes use of this function to determine
if a table has multiple (= more than one) elements. There's no need
for an 'exhaustive' search to achieve this, so we introduce a limit
(= max. number of elements) as optional parameter to CountTable().

The commit also adds some test cases for CountTable().